requirements.txt
torch==2.5.1
torchvision==0.20.1
ultralytics==8.3.58
opencv-python==4.10.0.84
python-telegram-bot==21.9
python-dotenv==1.0.1

